third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-release.html

<!DOCTYPE html>
<html>
    <head>
        <title>MediaKeySession lifetime after release()</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            async_test(function(test)
            {
                gc();
                var mediaKeys;
                var mediaKeySession1;
                var mediaKeySession2;
                var initDataType;
                var initData;
                var startingMediaKeysCount = window.internals.mediaKeysCount();
                var startingMediaKeySessionCount = window.internals.mediaKeySessionCount();

                function numMediaKeysCreated()
                {
                    return window.internals.mediaKeysCount() - startingMediaKeysCount;
                }

                function numMediaKeySessionCreated()
                {
                    return window.internals.mediaKeySessionCount() - startingMediaKeySessionCount;
                }

                // Create 2 sessions.
                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.create()');
                    assert_equals(numMediaKeySessionCreated(), 0, 'MediaKeys.create()');

                    mediaKeySession1 = mediaKeys.createSession();
                    return mediaKeySession1.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.createSession(1)');
                    assert_equals(numMediaKeySessionCreated(), 1, 'MediaKeys.createSession(1)');

                    mediaKeySession2 = mediaKeys.createSession();
                    return mediaKeySession2.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeys.createSession(2)');
                    assert_equals(numMediaKeySessionCreated(), 2, 'mediaKeys.createSession(2)');

                    // Close the sessions. Once completed, only the JS
                    // reference to them keeps them around.
                    return mediaKeySession1.close();
                }).then(function(result) {
                    return mediaKeySession2.close();
                }).then(function(result) {
                    // Since both sessions have been closed, dropping the
                    // reference to them from JS will result in the session
                    // being garbage-collected.
                    assert_equals(numMediaKeysCreated(), 1, 'after close');
                    assert_equals(numMediaKeySessionCreated(), 2, 'after close');

                    mediaKeySession1 = null;
                    return createGCPromise();
                }).then(function() {
                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeySession1 not collected');
                    assert_equals(numMediaKeySessionCreated(), 1, 'mediaKeySession1 not collected');

                    mediaKeySession2 = null;
                    return createGCPromise();
                }).then(function() {
                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeySession2 not collected');
                    assert_equals(numMediaKeySessionCreated(), 0, 'mediaKeySession2 not collected');
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeySession lifetime after release()');
        </script>
    </body>
</html>
